requirements.txt
dnspython==2.3.0
et-xmlfile==1.1.0
openpyxl==3.1.1
py4j==0.10.9.5
pymongo==4.3.3
pyspark==3.3.2
python-dotenv==0.21.1
